shop_user_service/requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
djangorestframework
Pillow
pytz
sqlparse
typing_extensions
pytest
pytest-django
requests
django-cors-headers
coreapi
coreschema
dj-rest-auth
django-allauth
djangorestframework-simplejwt
django-filter
httpie